+++ ci/generate-docs.py
@@ -343,6 +343,10 @@ TOC = [
                         "config/lua/wezterm.procinfo",
                     ),
                     Gen(
+                        "module: wezterm.time",
+                        "config/lua/wezterm.time",
+                    ),
+                    Gen(
                         "struct: Config",
                         "config/lua/config",
                     ),
@@ -360,6 +364,7 @@ TOC = [
                     Page("object: TabInformation", "config/lua/TabInformation.md"),
                     Page("object: SshDomain", "config/lua/SshDomain.md"),
                     Page("object: SpawnCommand", "config/lua/SpawnCommand.md"),
+                    Gen("object: Time", "config/lua/wezterm.time/Time"),
                     Page("object: TlsDomainClient", "config/lua/TlsDomainClient.md"),
                     Page("object: TlsDomainServer", "config/lua/TlsDomainServer.md"),
                     Gen(