+++ env-bootstrap/src/lib.rs
@@ -171,6 +171,7 @@ fn register_lua_modules() {
         ssh_funcs::register,
         spawn_funcs::register,
         share_data::register,
+        time_funcs::register,
     ] {
         config::lua::add_context_setup_func(func);
     }